requirements.txt
# ============================================================================
# EXPLAINABLE FRAUD DETECTION SYSTEM — PINNED DEPENDENCIES
# ============================================================================
# pip install -r requirements.txt
# ============================================================================

# Core
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
pyyaml==6.0.1
joblib==1.3.2

# Preprocessing & Imbalance
scikit-learn==1.4.1.post1
imbalanced-learn==0.12.0

# Models
xgboost==2.0.3
lightgbm==4.3.0

# Explainability
shap==0.45.0
lime==0.2.0.1

# Graph Analytics
networkx==3.2.1
pyvis==0.3.2

# API
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3

# Dashboard
streamlit==1.31.1
plotly==5.19.0

# Database (feedback loop)
sqlalchemy==2.0.27

# Visualization
matplotlib==3.8.3
seaborn==0.13.2

# Testing
pytest==8.0.2
pytest-cov==4.1.0
httpx==0.27.0

# Utilities
requests==2.31.0
